feat: Add a factory class responsible for creating CSTs
