+++ astars/core/cst/__init__.py
@@ -1,2 +1,3 @@
+from .factory import CSTFactory
 from .cst import CST
 from .node import CSTNode
